moving and renaming files in the abstract_graphs folder
