+++ graphlearn/abstract_graphs/rna/RNA.py
@@ -1,8 +1,8 @@
-from minortransform import MinorDecomposer
+from graphlearn.abstract_graphs.minordecompose import MinorDecomposer
 import eden
 import networkx as nx
 import subprocess as sp
-import forgi
+import graphlearn.abstract_graphs.forgi
 import eden.converter.rna as converter
 from eden import path
 import sklearn
@@ -170,9 +170,9 @@ class RnaDecomposer(MinorDecomposer):
         if self._abstract_graph is None:
 
             # create the abstract graph and populate the contracted set
-            abstract_graph = forgi.get_abstr_graph(self.structure, ignore_inserts=self.ignore_inserts)
+            abstract_graph = graphlearn.abstract_graphs.forgi.get_abstr_graph(self.structure, ignore_inserts=self.ignore_inserts)
             abstract_graph = self.vectorizer._edge_to_vertex_transform(abstract_graph)
-            self._abstract_graph = forgi.edge_parent_finder(abstract_graph, self._base_graph)
+            self._abstract_graph = graphlearn.abstract_graphs.forgi.edge_parent_finder(abstract_graph, self._base_graph)
 
             # eden is forcing us to set a label and a contracted attribute.. lets do this
             for n, d in self._abstract_graph.nodes(data=True):
